library, silkscreen and gerber update
I moved all the footprints into a single library so that if folks don't have the particular lib they won't have to go and search for it or try to replace. Silkscreen was updated again to show clearer positioning of components and gerbers updated to most recent. Version number isn't changing as no big updates were made.